# backend/internal/server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: GoMD API
  description: API for managing a self-hosted markdown knowledge base with real-time SSE updates and MCP integration.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Default local development server
paths:
  /api/files:
    get:
      summary: List all files
      responses:
        '200':
          description: Successful response
    post:
      summary: Create a new file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                path:
                  type: string
                  description: Path to the new file relative to vault root
                content:
                  type: string
                  description: File content
      responses:
        '201':
          description: File created
...
  /api/folders:
    get:
      summary: List root directory contents
      responses:
        '200':
          description: Returns files and directories
    post:
      summary: Create a new directory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                path:
                  type: string
      responses:
        '201':
          description: Directory created
...